+++ ITCC.UI/Windows/LogWindow.xaml
@@ -5,7 +5,8 @@
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
         xmlns:utils="clr-namespace:ITCC.UI.Utils"
         mc:Ignorable="d"
-        Title="Application log">
+        Title="Application log"
+        Loaded="LogWindow_OnLoaded">
     <Grid Name="MainGrid">
         <Grid.RowDefinitions>
             <RowDefinition Height="Auto" />
@@ -45,7 +46,6 @@
                   VerticalScrollBarVisibility="Visible"
                   EnableColumnVirtualization="True"
                   EnableRowVirtualization="True"
-                  IsReadOnly="True"
-                  utils:DataGridExtensions.AlwaysScrollToEnd="True" />
+                  IsReadOnly="True"/>
     </Grid>
 </Window>
